+++ src/ansys/aedt/core/application/analysis_3d.py
@@ -30,7 +30,6 @@ from typing import Union
 import warnings
 
 from ansys.aedt.core.application.analysis import Analysis
-from ansys.aedt.core.generic.checks import min_aedt_version
 from ansys.aedt.core.generic.configurations import Configurations
 from ansys.aedt.core.generic.constants import unit_converter
 from ansys.aedt.core.generic.file_utils import check_if_path_exists
@@ -40,6 +39,7 @@ from ansys.aedt.core.generic.file_utils import open_file
 from ansys.aedt.core.generic.file_utils import read_component_file
 from ansys.aedt.core.generic.general_methods import pyaedt_function_handler
 from ansys.aedt.core.generic.settings import settings
+from ansys.aedt.core.internal.checks import min_aedt_version
 
 
 class FieldAnalysis3D(Analysis, object):
